requirements.txt
Flask
flask-cors
Flask-SocketIO
python-socketio
requests
Werkzeug
Jinja2
python-dotenv
google-generativeai==0.3.2
google-genai==0.3.0
numpy
bcrypt==4.0.1
# Voice support dependencies
# SpeechRecognition==3.10.0
# pyaudio==0.2.11
# pyttsx3==2.90
# Payment dependencies
razorpay==1.2.0
